Decouple managed network services for plugins
Extract DHCPv4, RA, DHCPv6, IPv6 allocation, and runtime reconciliation into reusable packages. Add plugin network inventory APIs, fixed DHCP plans, unified effective network planning, and Linux integration coverage.

+++ internal/app/kernel_runtime_plugin_pipeline_linux_test.go
@@ -2324,7 +2324,7 @@ func TestKernelPluginPipelineRuntimeAttachesExplicitEgressInterfaceWithoutRules(
 	for _, attachment := range rt.attachments {
 		filter := attachment.filter
 		if filter.Parent != netlink.HANDLE_MIN_EGRESS {
-			t.Fatalf("filter parent = %#x, want HANDLE_MIN_EGRESS %#x", filter.Parent, netlink.HANDLE_MIN_EGRESS)
+			t.Fatalf("filter parent = %#x, want HANDLE_MIN_EGRESS %#x", filter.Parent, uint32(netlink.HANDLE_MIN_EGRESS))
 		}
 		if _, ok := wantNames[filter.Name]; !ok {
 			t.Fatalf("unexpected egress filter name %q", filter.Name)
